CoachSync/backend/requirements.txt
# requirements.txt for CoachSync backend
fastapi
uvicorn
celery
redis
python-dotenv
sqlalchemy[asyncio]
boto3
httpx
httpx
psycopg2-binary
asyncpg
PyJWT
python-multipart
requests
